% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDT.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set for firedam}
\format{
A data set of 15 rows and 2 columns
A data set of 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Distance in miles from a fire station}
  \item{DAMAGE}{Damage in Thousands of Dollars}

  ...
}
}
\source{
Text Book
}
\usage{
ddt
}
\description{
15 objects with 2 variables pertaining to fire damage
Report ...
}
\keyword{datasets}
