% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{Title}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{Initial guess for the parameter of interest.}

\item{delta}{Small increment used to approximate the derivative.}

\item{llik}{Function computing the log-likelihood.}

\item{xrange}{Range to be considered for plotting (not currently used in the function).}

\item{parameter}{Name of the parameter to be estimated.}
}
\value{
List containing the estimated parameter value and the number of iterations.
}
\description{
Title
}
\examples{
# Set up a simple log-likelihood function
llik <- function(lambda) { sum(dpois(x=4:10, lambda=lambda, log=TRUE)) }
# Find the maximum likelihood estimate for lambda starting from lambda=1
myNRML(x0=1, llik=llik, xrange=c(0,10), parameter="lambda")
}
