% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myquad.R
\name{myquad}
\alias{myquad}
\title{Quadratic Function}
\usage{
myquad(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector representing the result of the quadratic function.
}
\description{
This function calculates the quadratic function for a given numeric vector.
}
\examples{
x <- 1:10
myquad(x)

}
\seealso{
Use \code{\link{plot}} for visualizing the quadratic function.
}
\keyword{internal}
