% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculate the number of tickets to be sold}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats in the flight.}

\item{gamma}{Probability a plane will be truly overbooked (more people show than there are seats).}

\item{p}{Probability of a "show".}
}
\value{
A list containing nd, nc, N, p, and gamma.
}
\description{
This function calculates the number of tickets to be sold for a flight given
the number of seats, the probability of a "show", and the probability of overbooking.
It provides two methods of calculation: one using a discrete distribution and
the other using a normal approximation.
}
\details{
The function calculates nd using the appropriate discrete distribution
and nc using the normal approximation. It then prints a named list containing
nd, nc, N, p, and gamma. Additionally, it creates plots of the objective function
versus n for both the discrete and continuous cases.
}
\examples{
ntickets(N = 150, gamma = 0.1, p = 0.05)
}
