% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Z score from a vector
Compute Z score for outlier detection
Input parameter must be a vector}
\usage{
zscore(x)
}
\arguments{
\item{x}{a vector of interest}
}
\value{
z score of a vector
}
\description{
Z score from a vector
Compute Z score for outlier detection
Input parameter must be a vector
}
\examples{
zscore(x)
}
